feat: add feature to configure non internet facing
BREAKING CHANGE: This commit needs to set a new variable to configure
the visibility of the service.
If the Service is internet facing you should configure the
WF_INTERNET_FACING: TRUE, Otherwise , its value is false.

footer: This commit is needed to set the visibility of the services on
internet facing or local